+++ docs/html/search/variables_4.js
@@ -1,7 +1,8 @@
 var searchData=
 [
-  ['fold_428',['fold',['../structAtcLocalDateTime.html#a51a2be455f91b1aa97263263ac200fe4',1,'AtcLocalDateTime::fold()'],['../structAtcOffsetDateTime.html#a6ede44821779979e479a7d19718dab81',1,'AtcOffsetDateTime::fold()'],['../structAtcTransitionForSeconds.html#a151459cc53e494c06b61a8cb679077f4',1,'AtcTransitionForSeconds::fold()'],['../structAtcFindResult.html#ac2e538e6cc5c37537082931893d5f797',1,'AtcFindResult::fold()'],['../structAtcZonedDateTime.html#aec3c3d0be1b2fa27bd91752b4b0dbcb8',1,'AtcZonedDateTime::fold()']]],
-  ['format_429',['format',['../structAtcZoneEra.html#a4242f6211236aa265ea77d8a56e96bd6',1,'AtcZoneEra']]],
-  ['fragments_430',['fragments',['../structAtcZoneContext.html#a4b06c0da451ac5578b5a251e8902166f',1,'AtcZoneContext']]],
-  ['from_5fyear_431',['from_year',['../structAtcZoneRule.html#a3d0e03f6dc2138ec65e68223128912d5',1,'AtcZoneRule']]]
+  ['fold_455',['fold',['../structAtcLocalDateTime.html#a51a2be455f91b1aa97263263ac200fe4',1,'AtcLocalDateTime::fold()'],['../structAtcOffsetDateTime.html#a6ede44821779979e479a7d19718dab81',1,'AtcOffsetDateTime::fold()'],['../structAtcTransitionForSeconds.html#a151459cc53e494c06b61a8cb679077f4',1,'AtcTransitionForSeconds::fold()'],['../structAtcFindResult.html#ac2e538e6cc5c37537082931893d5f797',1,'AtcFindResult::fold()'],['../structAtcZonedDateTime.html#aec3c3d0be1b2fa27bd91752b4b0dbcb8',1,'AtcZonedDateTime::fold()']]],
+  ['fold_5ftype_456',['fold_type',['../structAtcZonedExtra.html#aaff1381eaedc5e5885d6b9bb1ae19e72',1,'AtcZonedExtra']]],
+  ['format_457',['format',['../structAtcZoneEra.html#a4242f6211236aa265ea77d8a56e96bd6',1,'AtcZoneEra']]],
+  ['fragments_458',['fragments',['../structAtcZoneContext.html#a4b06c0da451ac5578b5a251e8902166f',1,'AtcZoneContext']]],
+  ['from_5fyear_459',['from_year',['../structAtcZoneRule.html#a3d0e03f6dc2138ec65e68223128912d5',1,'AtcZoneRule']]]
 ];