+++ src/acetimec/local_date.c
@@ -75,11 +75,10 @@ uint8_t atc_local_date_day_of_week(int16_t year, uint8_t month, uint8_t day)
   return (d < -1) ? (d + 1) % 7 + 8 : (d + 1) % 7 + 1;
 }
 
-int32_t atc_local_date_to_epoch_days(
-    int16_t year, uint8_t month, uint8_t day)
+int32_t atc_local_date_to_epoch_days(int16_t year, uint8_t month, uint8_t day)
 {
-  int32_t converter_days = atc_convert_to_days(year, month, day);
-  return converter_days - atc_get_days_to_current_epoch_from_converter_epoch();
+  int32_t internal_days = atc_convert_to_internal_days(year, month, day);
+  return internal_days - atc_days_to_current_epoch_from_internal_epoch;
 }
 
 void atc_local_date_from_epoch_days(
@@ -88,9 +87,25 @@ void atc_local_date_from_epoch_days(
     uint8_t *month,
     uint8_t *day)
 {
-  int32_t converter_days = epoch_days
-      + atc_get_days_to_current_epoch_from_converter_epoch();
-  atc_convert_from_days(converter_days, year, month, day);
+  int32_t internal_days = epoch_days
+      + atc_days_to_current_epoch_from_internal_epoch;
+  atc_convert_from_internal_days(internal_days, year, month, day);
+}
+
+int32_t atc_local_date_to_unix_days(int16_t year, uint8_t month, uint8_t day)
+{
+  int32_t internal_days = atc_convert_to_internal_days(year, month, day);
+  return internal_days + kAtcDaysToInternalEpochFromUnixEpoch;
+}
+
+void atc_local_date_from_unix_days(
+    int32_t unix_days,
+    int16_t *year,
+    uint8_t *month,
+    uint8_t *day)
+{
+  int32_t internal_days = unix_days - kAtcDaysToInternalEpochFromUnixEpoch;
+  atc_convert_from_internal_days(internal_days, year, month, day);
 }
 
 void atc_local_date_increment_one_day(