+++ src/acetimec/local_date_time.c
@@ -6,6 +6,7 @@
 #include "local_date_time.h"
 
 #include "common.h"
+#include "epoch.h"
 #include "local_date.h"
 #include "local_time.h"
 #include "local_date_time.h"
@@ -23,7 +24,7 @@ bool atc_local_date_time_is_error(const AtcLocalDateTime *ldt)
 atc_time_t atc_local_date_time_to_epoch_seconds(
     const AtcLocalDateTime *ldt)
 {
-  if (ldt->year == kAtcInvalidYear) return kAtcInvalidEpochSeconds;
+  if (atc_local_date_time_is_error(ldt)) return kAtcInvalidEpochSeconds;
 
   int32_t days = atc_local_date_to_epoch_days(
       ldt->year, ldt->month, ldt->day);
@@ -58,8 +59,50 @@ void atc_local_date_time_from_epoch_seconds(
   uint16_t minutes = seconds / 60;
   ldt->minute = minutes % 60;
   ldt->hour = minutes / 60;
+}
 
-  return;
+int64_t atc_local_date_time_to_unix_seconds(const AtcLocalDateTime *ldt) {
+  if (atc_local_date_time_is_error(ldt)) return kAtcInvalidUnixSeconds;
+
+  int32_t days = atc_local_date_to_epoch_days(
+      ldt->year, ldt->month, ldt->day);
+  int32_t seconds = atc_local_time_to_seconds(
+      ldt->hour, ldt->minute, ldt->second);
+  int32_t unix_days = atc_unix_days_from_epoch_days(days);
+  return unix_days * (int64_t)86400 + seconds;
+}
+
+void atc_local_date_time_from_unix_seconds(
+  AtcLocalDateTime *ldt,
+  int64_t unix_seconds)
+{
+  ldt->fold = 0;
+
+  if (unix_seconds == kAtcInvalidUnixSeconds) {
+    atc_local_date_time_set_error(ldt);
+    return;
+  }
+
+  // Integer floor-division towards -infinity. Implicitly assume that
+  // unix_seconds/86400 fits inside an int32_t. So the largest/smallest days is
+  // about +/- 2^31, which translates to +/- 5_879_489 years, which I think is
+  // more than sufficient for the foreseeable future.
+  int32_t unix_days = (unix_seconds < 0)
+      ? (unix_seconds + 1) / 86400 - 1
+      : unix_seconds / 86400;
+  int32_t seconds = unix_seconds - 86400 * unix_days;
+
+  int32_t days = atc_epoch_days_from_unix_days(unix_days);
+
+  // Extract (year, month day).
+  atc_local_date_from_epoch_days(days, &ldt->year, &ldt->month, &ldt->day);
+
+  // Extract (hour, minute, second). The compiler will combine the mod (%) and
+  // division (/) operations into a single (dividend, remainder) function call.
+  ldt->second = seconds % 60;
+  uint16_t minutes = seconds / 60;
+  ldt->minute = minutes % 60;
+  ldt->hour = minutes / 60;
 }
 
 void atc_local_date_time_print(