+++ src/acetimec/zoned_extra.c
@@ -3,17 +3,18 @@
  * Copyright (c) 2022 Brian T. Park
  */
 
+#include "epoch.h"
 #include "time_zone.h"
 #include "zoned_extra.h"
 
 void atc_zoned_extra_set_error(AtcZonedExtra *extra)
 {
-  extra->type = kAtcZonedExtraNotFound;
+  extra->fold_type = kAtcFoldTypeNotFound;
 }
 
 bool atc_zoned_extra_is_error(const AtcZonedExtra *extra)
 {
-  return extra->type == kAtcZonedExtraNotFound;
+  return extra->fold_type == kAtcFoldTypeNotFound;
 }
 
 void atc_zoned_extra_from_epoch_seconds(
@@ -21,8 +22,16 @@ void atc_zoned_extra_from_epoch_seconds(
     atc_time_t epoch_seconds,
     const AtcTimeZone *tz)
 {
-  atc_time_zone_zoned_extra_from_epoch_seconds(
-      tz, epoch_seconds, extra);
+  atc_time_zone_zoned_extra_from_epoch_seconds(tz, epoch_seconds, extra);
+}
+
+void atc_zoned_extra_from_unix_seconds(
+    AtcZonedExtra *extra,
+    int64_t unix_seconds,
+    const AtcTimeZone *tz)
+{
+  atc_time_t epoch_seconds = atc_epoch_seconds_from_unix_seconds(unix_seconds);
+  atc_time_zone_zoned_extra_from_epoch_seconds(tz, epoch_seconds, extra);
 }
 
 void atc_zoned_extra_from_local_date_time(