+++ src/acetimec/zoned_extra.h
@@ -27,22 +27,28 @@ typedef struct AtcTimeZone AtcTimeZone;
 typedef struct AtcLocalDateTime AtcLocalDateTime;
 
 /**
- * Values of the the AtcZonedExtra.type field. Must be identical to the
+ * Values of the the AtcZonedExtra.fold_type field. Must be identical to the
  * corresponding AtcFindResultXxx enums in zone_processor.h.
  */
 enum {
-  kAtcZonedExtraNotFound = 0, // rename this to kAtcZonedExtraError?
-  kAtcZonedExtraExact = 1,
-  kAtcZonedExtraOverlap = 2,
-  kAtcZonedExtraGap = 3,
+  kAtcFoldTypeNotFound = 0, // rename this to kAtcFoldTypeError?
+  kAtcFoldTypeExact = 1,
+  kAtcFoldTypeOverlap = 2,
+  kAtcFoldTypeGap = 3,
 };
 
 /**
  * Extra information about a given time zone at a specified epoch seconds.
  */
 typedef struct AtcZonedExtra {
-  /** Type of match against the LocalDateTime or epoch_seconds. */
-  int8_t type;
+  /**
+   * Result of search for the LocalDateTime or epoch_seconds, which is
+   * determines by the type of fold.
+   */
+  int8_t fold_type;
+
+  /** abbreviation (e.g. PST, PDT) */
+  char abbrev[kAtcAbbrevSize];
 
   /** STD offset */
   int32_t std_offset_seconds;
@@ -55,20 +61,22 @@ typedef struct AtcZonedExtra {
 
   /** DST offset of the requested LocalDateTime or epoch_seconds */
   int32_t req_dst_offset_seconds;
-
-  /** abbreviation (e.g. PST, PDT) */
-  char abbrev[kAtcAbbrevSize]; // TODO: Move this after 'type' to save space
 } AtcZonedExtra;
 
-/** Set the given AtcZonedDateTime to its error state, i.e. NotFound. */
+/**
+ * Set the given AtcZonedDateTime to its error state, i.e. kFoldTypeNotFound.
+ */
 void atc_zoned_extra_set_error(AtcZonedExtra *extra);
 
-/** Return true if AtcZonedExtra is an error. */
+/**
+ * Return true if AtcZonedExtra is an error defined by fold_type ==
+ * kFoldTypeNotFound.
+ */
 bool atc_zoned_extra_is_error(const AtcZonedExtra *extra);
 
 /**
- * Extract the extra zone information at given epoch_seconds.
- * Returns error status in `extra->type`.
+ * Extract the extra zone information at given epoch seconds.
+ * Returns error status in `extra->fold_type`.
  */
 void atc_zoned_extra_from_epoch_seconds(
     AtcZonedExtra *extra,
@@ -76,8 +84,21 @@ void atc_zoned_extra_from_epoch_seconds(
     const AtcTimeZone *tz);
 
 /**
+ * Extract the extra zone information at given Unix seconds. Since this uses
+ * the AtcZoneProcessor and its transition cache, the range of supported Unix
+ * seconds is determined by the range of the epoch seconds in the context of the
+ * current epoch year.
+ *
+ * Returns `extra->fold_type == kAtcFoldTypeNotFound` if not found.
+ */
+void atc_zoned_extra_from_unix_seconds(
+    AtcZonedExtra *extra,
+    int64_t unix_seconds,
+    const AtcTimeZone *tz);
+
+/**
  * Extract the extra zone information at given LocalDateTime.
- * Returns error status in `extra->type`.
+ * Returns `extra->fold_type == kAtcFoldTypeNotFound` if not found.
  */
 void atc_zoned_extra_from_local_date_time(
     AtcZonedExtra *extra,